Move ReactionsList components to separate files
